requirements.txt
markdown-it-py[linkify,plugins]
readability-lxml
markdownify
requests